refacto(web): PlanContext et éclatement de PlanSidebar
`PlanSidebar` prenait 33 props, étalées depuis `PlanPage` dans deux rendus
qu'il fallait garder en phase à la main, et portait neuf sous-composants dans
1253 lignes (annexe B, C2). Presque aucune de ces props ne lui appartenait :
c'était la session, enfilée à travers des composants qui ne faisaient que la
transmettre.

`plan/session/planContext.ts` publie `{state, actions}` plus les trois choses
que la session ne possède pas : le catalogue de bateaux, le spinner du mode à
l'écran, et les deux calculs enveloppés pour recadrer la carte. Le provider
vit dans son propre fichier pour que le module de contexte reste sans JSX.

Un fichier par sous-composant sous `plan/sidebar/`, plus une découpe des trois
vues remplies (`PlanForm`, `CompareResults`, `SingleResults`). `PlanSidebar`
ne garde que l'aiguillage entre les cinq vues, dont l'ordre est porteur :
1253 → 121 lignes.

`aggregateLegs` est mémoïsé dans `LegList` : il parcourait tous les tronçons
depuis une IIFE en JSX, donc à chaque tick du curseur de départ.

`config/usePolarConfig.ts` remplace les relectures par render : un
`useSyncExternalStore` sur deux canaux, le registre de module de
`polarConfig.ts` pour cet onglet et l'événement `storage` pour les autres. Le
navigateur ne déclenche `storage` que pour les autres onglets, donc le bateau
choisi sur /plan n'aurait jamais atteint la bande de récap sans le registre.
L'instantané est mémoïsé sur la chaîne brute.

18 tests ajoutés (12 sur les cinq vues du panneau montées pour de vrai, 6 sur
le nouveau store). PlanSidebar.tsx : 1253 → 121 lignes. PlanPage.tsx :
682 → 633. Budget lint inchangé à 6.
