+++ packages/web/src/plan/PlanSidebar.tsx
@@ -1,1253 +1,121 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 Quentin Donnars
 
-import { useCallback, useMemo, useState } from "react";
-import { useTheme } from "../design/useTheme";
-import { useBackDismiss } from "../hooks/useBackDismiss";
+/**
+ * The planner panel: which of its five views is on screen, and nothing else.
+ *
+ * This file used to be 1253 lines and take 33 props, with nine sub-components
+ * inlined (annexe B, C2). Each of them now lives in its own file under
+ * `sidebar/` and reads the session from `PlanContext`, so what is left here is
+ * the branch itself.
+ *
+ * Its order is load-bearing: every view assumes the ones above it did not
+ * match.
+ *
+ * 1. computing: a skeleton, so the panel never shows half a plan;
+ * 2. failed: the error, with the mode pills still reachable;
+ * 3. no route yet: the empty state;
+ * 4. a route but no mode picked: the compact pick-a-mode step;
+ * 5. the picked mode, filled when it has a result, as a form otherwise.
+ */
+
 import { validateSweep, type SweepValidation } from "./validateSweep";
-import type { PassageReport, ComplexityScore, Archetype, PassageWindow } from "./types";
-import { aggregateLegs, buildLegSummaryCells, type AggregatedLeg } from "./aggregateLegs";
-import { cxLevel, CX_COLORS } from "./types";
-import { WindowsTable } from "./WindowsTable";
-import { ModeToggle, TimeAnchorToggle, type PlanMode, type TimeAnchor } from "./ModeToggle";
-import { LegDetailCard } from "./LegDetailCard";
-import { EmptyState, ModePicker, Warn, RecapButton, HeroStats } from "./PlanStates";
-import {
-  ARCHETYPE_LABELS,
-  defaultPolarConfig,
-  isImportedActive,
-  isPersoActive,
-  isPolarCustomized,
-  loadPolarConfig,
-  planMinUpwind,
-  savePolarConfig,
-} from "../config/polarConfig";
-import { rememberReturnPath } from "../config/returnPath";
+import { EmptyState, ModePicker } from "./PlanStates";
+import { usePolarConfig } from "../config/usePolarConfig";
+import { usePlan } from "./session/planContext";
+import { PlanHeaderRow } from "./sidebar/PlanHeaderRow";
+import { PlanForm } from "./sidebar/PlanForm";
+import { CompareResults } from "./sidebar/CompareResults";
+import { SingleResults } from "./sidebar/SingleResults";
+import { boatLabel } from "./sidebar/boatLabel";
 
-// ── helpers ──────────────────────────────────────────────────────────────────
-
-// Format a Date as "YYYY-MM-DDTHH:MM" (local, naive — the format datetime-local expects).
-function toLocalIso(d: Date): string {
-  const pad = (n: number) => String(n).padStart(2, "0");
-  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}T${pad(d.getHours())}:${pad(d.getMinutes())}`;
-}
-
-// ── DepartureSlider ──────────────────────────────────────────────────────────
-
-// Open-Meteo's forecast endpoint caps start_date/end_date at ~today+15. We cap
-// the slider at 14 d to leave 1 d of margin (clock skew, TZ crossings).
-const SLIDER_MAX_HOURS = 14 * 24;
-
-function DepartureSlider({
-  value,
-  onChange,
-  resolvedTheme,
-  anchor: timeAnchor = "departure",
-}: {
-  value: string;
-  onChange: (iso: string) => void;
-  resolvedTheme: "light" | "dark";
-  anchor?: TimeAnchor;
-}) {
-  const [showManual, setShowManual] = useState(false);
-
-  // Anchor "now" once per mount so the slider's left edge stays fixed during interaction.
-  const anchor = useMemo(() => {
-    const d = new Date();
-    d.setMinutes(0, 0, 0);
-    return d;
-  }, []);
-
-  const valueDate = useMemo(() => new Date(value), [value]);
-  const valueHours = Math.max(
-    0,
-    Math.min(SLIDER_MAX_HOURS, Math.round((valueDate.getTime() - anchor.getTime()) / 3_600_000)),
-  );
-
-  function setHours(h: number) {
-    const d = new Date(anchor.getTime() + h * 3_600_000);
-    onChange(toLocalIso(d));
-  }
-
-  function resetToNow() {
-    const d = new Date();
-    d.setMinutes(Math.ceil(d.getMinutes() / 15) * 15, 0, 0);
-    onChange(toLocalIso(d));
-  }
-
-  // Display labels — section header changes with the time anchor.
-  const sectionLabel = timeAnchor === "arrival" ? "Arrivée" : "Départ";
-  const ariaLabel = timeAnchor === "arrival" ? "Heure d'arrivée souhaitée" : "Date de départ";
-  const dateLabel = valueDate.toLocaleDateString("fr-FR", { weekday: "long", day: "numeric", month: "long" });
-  const timeLabel = valueDate.toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" });
-  const dayDelta = Math.floor((valueDate.getTime() - anchor.getTime()) / 86_400_000);
-  const offsetLabel =
-    dayDelta <= 0 ? "Aujourd'hui" :
-    dayDelta === 1 ? "Demain" :
-    `Dans ${dayDelta} jours`;
-
+/** Placeholder while a computation runs, so the panel never shows half a plan. */
+function LoadingSkeleton() {
   return (
-    <div>
-      <div className="flex items-baseline justify-between mb-1">
-        <span className="text-[10px] uppercase tracking-widest font-semibold" style={{ color: "var(--ow-fg-2)" }}>
-          {sectionLabel}
-        </span>
-        <button
-          type="button"
-          onClick={() => setShowManual((v) => !v)}
-          className="text-[10px] underline"
-          style={{ color: "var(--ow-fg-2)" }}
-        >
-          {showManual ? "Slider" : "Ajuster"}
-        </button>
+    <div className="p-4 space-y-4 animate-fade-in">
+      <div className="skeleton h-8 w-48 rounded-lg" />
+      <div className="skeleton h-5 w-32 rounded" />
+      <div className="grid grid-cols-4 gap-2">
+        {[0, 1, 2, 3].map((i) => <div key={i} className="skeleton h-14 rounded-lg" />)}
       </div>
-
-      <div className="text-sm font-semibold mb-2" style={{ color: "var(--ow-fg-0)" }}>
-        <span className="capitalize">{dateLabel}</span>
-        <span className="mx-1.5" style={{ color: "var(--ow-fg-2)" }}>·</span>
-        <span className="tabular-nums" style={{ fontFamily: "var(--ow-font-mono)" }}>{timeLabel}</span>
-        <span className="ml-2 text-[11px] font-normal" style={{ color: "var(--ow-fg-2)" }}>{offsetLabel}</span>
-      </div>
-
-      {showManual ? (
-        <input
-          type="datetime-local"
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          className="ow-datetime-input w-full rounded-lg px-3 py-2 text-sm font-semibold tabular-nums"
-          style={{
-            background: "var(--ow-bg-2)",
-            color: "var(--ow-fg-0)",
-            border: "1px solid var(--ow-line)",
-            fontFamily: "var(--ow-font-mono)",
-            colorScheme: resolvedTheme === "light" ? "light" : "dark",
-          }}
-        />
-      ) : (
-        <>
-          <input
-            type="range"
-            min={0}
-            max={SLIDER_MAX_HOURS}
-            step={1}
-            value={valueHours}
-            onChange={(e) => setHours(Number(e.target.value))}
-            className="ow-departure-slider w-full"
-            aria-label={ariaLabel}
-          />
-          <div className="flex justify-between text-[10px] mt-1" style={{ color: "var(--ow-fg-2)" }}>
-            {/* Maintenant is clickable: defaulting to J+1 lets the user pick a
-                horizon, but a single tap still lands them back at "now". */}
-            <button
-              type="button"
-              onClick={resetToNow}
-              className="underline-offset-2 hover:underline"
-              style={{ color: dayDelta <= 0 ? "var(--ow-accent)" : "var(--ow-fg-2)" }}
-            >
-              Maintenant
-            </button>
-            <span>+1 sem.</span>
-            <span>+2 sem.</span>
-          </div>
-        </>
-      )}
+      <div className="skeleton h-2.5 rounded-full" />
+      {[0, 1, 2].map((i) => <div key={i} className="skeleton h-10 rounded-lg" />)}
     </div>
   );
 }
 
+export function PlanSidebar() {
+  const { state, actions, archetypes, isLoading } = usePlan();
+  const {
+    passage,
+    complexity,
+    windows,
+    apiError,
+    mode,
+    sweepEarliest,
+    sweepLatest,
+    sweepIntervalHours,
+    actionTaken,
+    waypoints,
+    archetype,
+  } = state;
+  const polarConfig = usePolarConfig();
 
+  const sweepValid: SweepValidation =
+    mode === "compare"
+      ? validateSweep(sweepEarliest, sweepLatest, sweepIntervalHours)
+      : { ok: true };
+  const canCalculate = waypoints.length >= 2 && (mode === "single" || sweepValid.ok);
 
-// ModeToggle moved to ./ModeToggle so CompactDrawer can reuse it.
+  // 1. computing
+  if (isLoading) return <LoadingSkeleton />;
 
-// ── DepartureRangeSlider ─────────────────────────────────────────────────────
-// Dual-thumb slider for the compare-windows form. Two overlapping native ranges
-// keyed in hours-from-now; the right thumb is constrained to stay >= 1 h after
-// the left thumb. Constraints by construction = no out-of-range dates can be
-// picked, replacing the error-prone datetime-local pair.
-
-const RANGE_MAX_HOURS = 14 * 24; // mirror Open-Meteo cap (today+14d)
-
-function DepartureRangeSlider({
-  earliestHours,
-  latestHours,
-  onChange,
-}: {
-  earliestHours: number;
-  latestHours: number;
-  onChange: (earliest: number, latest: number) => void;
-}) {
-  const anchor = useMemo(() => {
-    const d = new Date();
-    d.setMinutes(0, 0, 0);
-    return d;
-  }, []);
-
-  const eClamped = Math.max(0, Math.min(RANGE_MAX_HOURS - 1, earliestHours));
-  const lClamped = Math.max(eClamped + 1, Math.min(RANGE_MAX_HOURS, latestHours));
-
-  function setEarliest(h: number) {
-    const next = Math.max(0, Math.min(RANGE_MAX_HOURS - 1, h));
-    const newLatest = Math.max(next + 1, lClamped);
-    onChange(next, Math.min(RANGE_MAX_HOURS, newLatest));
-  }
-  function setLatest(h: number) {
-    const next = Math.min(RANGE_MAX_HOURS, Math.max(eClamped + 1, h));
-    onChange(eClamped, next);
-  }
-
-  function fmt(hours: number): { date: string; time: string; offset: string } {
-    const d = new Date(anchor.getTime() + hours * 3_600_000);
-    return {
-      date: d.toLocaleDateString("fr-FR", { weekday: "short", day: "numeric", month: "short" }),
-      time: d.toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" }),
-      offset: (() => {
-        const days = Math.floor(hours / 24);
-        if (days === 0) return "Aujourd'hui";
-        if (days === 1) return "Demain";
-        return `J+${days}`;
-      })(),
-    };
-  }
-
-  const eFmt = fmt(eClamped);
-  const lFmt = fmt(lClamped);
-  const windowHours = lClamped - eClamped;
-  const windowLabel = windowHours >= 24
-    ? `${(windowHours / 24).toFixed(windowHours % 24 === 0 ? 0 : 1)} j`
-    : `${windowHours} h`;
-
-  // Track fill % for visual fill between thumbs
-  const fillStart = (eClamped / RANGE_MAX_HOURS) * 100;
-  const fillEnd = (lClamped / RANGE_MAX_HOURS) * 100;
-
-  return (
-    <div>
-      <div className="flex items-baseline justify-between mb-1">
-        <span className="text-[10px] uppercase tracking-widest font-semibold" style={{ color: "var(--ow-fg-2)" }}>
-          Fenêtre de départ
-        </span>
-        <span className="text-[10px] tabular-nums" style={{ color: "var(--ow-accent)" }}>
-          {windowLabel}
-        </span>
-      </div>
-
-      {/* Two readouts, one per thumb */}
-      <div className="flex items-baseline justify-between text-xs mb-2 tabular-nums" style={{ fontFamily: "var(--ow-font-mono)" }}>
-        <div>
-          <div className="capitalize font-semibold" style={{ color: "var(--ow-fg-0)" }}>
-            {eFmt.date} · {eFmt.time}
-          </div>
-          <div className="text-[10px]" style={{ color: "var(--ow-fg-2)" }}>{eFmt.offset}</div>
-        </div>
-        <div className="text-right">
-          <div className="capitalize font-semibold" style={{ color: "var(--ow-fg-0)" }}>
-            {lFmt.date} · {lFmt.time}
-          </div>
-          <div className="text-[10px]" style={{ color: "var(--ow-fg-2)" }}>{lFmt.offset}</div>
-        </div>
-      </div>
-
-      <div className="ow-range-track">
-        <div className="ow-range-track-bg" />
-        <div
-          className="ow-range-track-fill"
-          style={{ left: `${fillStart}%`, right: `${100 - fillEnd}%` }}
-        />
-        <input
-          type="range"
-          min={0}
-          max={RANGE_MAX_HOURS}
-          step={1}
-          value={eClamped}
-          onChange={(e) => setEarliest(Number(e.target.value))}
-          className="ow-range-input"
-          aria-label="Départ au plus tôt"
-          style={{ zIndex: eClamped > RANGE_MAX_HOURS / 2 ? 3 : 2 }}
-        />
-        <input
-          type="range"
-          min={0}
-          max={RANGE_MAX_HOURS}
-          step={1}
-          value={lClamped}
-          onChange={(e) => setLatest(Number(e.target.value))}
-          className="ow-range-input"
-          aria-label="Départ au plus tard"
-          style={{ zIndex: eClamped > RANGE_MAX_HOURS / 2 ? 2 : 3 }}
-        />
-      </div>
-
-      <div className="flex justify-between text-[10px] mt-1" style={{ color: "var(--ow-fg-2)" }}>
-        <span>Maintenant</span>
-        <span>+1 sem.</span>
-        <span>+2 sem.</span>
-      </div>
-    </div>
-  );
-}
-
-// ── SweepForm ─────────────────────────────────────────────────────────────────
-
-const SWEEP_INTERVALS: { value: number; label: string }[] = [
-  { value: 1, label: "Toutes les heures" },
-  { value: 3, label: "Toutes les 3h" },
-  { value: 6, label: "Toutes les 6h" },
-];
-
-// Minimal "YYYY-MM-DDTHH:MM" formatter from a Date (local time).
-function toLocalIsoMin(d: Date): string {
-  const pad = (n: number) => String(n).padStart(2, "0");
-  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}T${pad(d.getHours())}:${pad(d.getMinutes())}`;
-}
-
-function SweepForm({
-  earliest,
-  latest,
-  intervalHours,
-  onEarliestChange,
-  onLatestChange,
-  onIntervalChange,
-}: {
-  earliest: string;
-  latest: string;
-  intervalHours: number;
-  onEarliestChange: (iso: string) => void;
-  onLatestChange: (iso: string) => void;
-  onIntervalChange: (h: number) => void;
-}) {
-  // Convert ISO local strings <-> hours-from-now so the slider can drive them.
-  const anchor = useMemo(() => {
-    const d = new Date();
-    d.setMinutes(0, 0, 0);
-    return d;
-  }, []);
-  function isoToHours(iso: string): number {
-    if (!iso) return 0;
-    const t = new Date(iso).getTime();
-    if (Number.isNaN(t)) return 0;
-    return Math.round((t - anchor.getTime()) / 3_600_000);
-  }
-  function hoursToIso(h: number): string {
-    return toLocalIsoMin(new Date(anchor.getTime() + h * 3_600_000));
-  }
-
-  const earliestHours = Math.max(0, isoToHours(earliest));
-  const latestHours = Math.max(earliestHours + 1, isoToHours(latest));
-
-  const validation = validateSweep(earliest, latest, intervalHours);
-
-  return (
-    <div className="space-y-3">
-      <DepartureRangeSlider
-        earliestHours={earliestHours}
-        latestHours={latestHours}
-        onChange={(e, l) => {
-          onEarliestChange(hoursToIso(e));
-          onLatestChange(hoursToIso(l));
-        }}
-      />
-      {!validation.ok && validation.message && (
-        <p className="text-[11px]" style={{ color: "var(--ow-warn)" }}>
-          {validation.message}
-        </p>
-      )}
-
-      <div>
-        <label className="block text-[10px] uppercase tracking-widest font-semibold mb-1" style={{ color: "var(--ow-fg-2)" }}>
-          Pas d'échantillonnage
-        </label>
-        <div className="flex gap-1.5">
-          {SWEEP_INTERVALS.map(({ value, label }) => {
-            const active = intervalHours === value;
-            return (
-              <button
-                key={value}
-                onClick={() => onIntervalChange(value)}
-                className="flex-1 px-2 py-1.5 rounded-lg text-xs font-semibold transition-colors"
-                style={{
-                  background: active ? "var(--ow-accent-soft)" : "var(--ow-bg-2)",
-                  color: active ? "var(--ow-accent)" : "var(--ow-fg-1)",
-                  border: `1px solid ${active ? "var(--ow-accent)" : "var(--ow-line-2)"}`,
-                }}
-              >
-                {label}
-              </button>
-            );
-          })}
-        </div>
-      </div>
-
-    </div>
-  );
-}
-
-// ── ArchetypeSelector ─────────────────────────────────────────────────────────
-
-function ArchetypeSelector({
-  currentSlug,
-  archetypes,
-  onChange,
-  onPersoSelect,
-  onCustomCleared,
-}: {
-  currentSlug: string;
-  archetypes: Archetype[];
-  onChange: (slug: string) => void;
-  // Select the « Perso » entry (shown first as soon as the polar deviates
-  // from the stock archetype default).
-  onPersoSelect: () => void;
-  // Notify the parent that the custom polar was reset so it can re-fetch
-  // (the new fingerprint won't match the cached one, and resolveOverrides
-  // will stop sending the custom matrix).
-  onCustomCleared?: () => void;
-}) {
-  const [open, setOpen] = useState(false);
-  // Back closes the boat picker instead of leaving the planner (issue #300),
-  // like the click-away overlay below.
-  const close = useCallback(() => setOpen(false), []);
-  useBackDismiss(open, close);
-  // Re-read the polar config every render. Cheap (localStorage parse + a few
-  // boolean checks) and ensures the « Perso » entry appears immediately after
-  // the user edits /config in another tab and comes back here.
-  const polarCfg = loadPolarConfig();
-  const persoDefined = isPolarCustomized(polarCfg);
-  const persoSelected = isPersoActive(polarCfg);
-  const baseLabel = ARCHETYPE_LABELS[polarCfg.base] ?? polarCfg.base;
-  const current = archetypes.find((a) => a.slug === currentSlug);
-  const label = persoSelected ? "Perso" : (current?.name ?? currentSlug);
-  // Detail line of the « Perso » entry, same shape as the archetype rows:
-  // provenance of the customization rather than hull specs.
-  const persoDetail = isImportedActive(polarCfg)
-    ? `${polarCfg.imported?.name ?? "polaire"} · importée`
-    : `${baseLabel} · ajustée`;
-
-  function resetCustom() {
-    savePolarConfig(defaultPolarConfig());
-    setOpen(false);
-    onCustomCleared?.();
-  }
-
-  return (
-    <div className="relative">
-      <button
-        onClick={() => setOpen((v) => !v)}
-        className="flex items-center gap-1.5 text-sm transition-colors"
-        style={{ color: persoSelected ? "var(--ow-accent)" : "var(--ow-fg-1)" }}
-        title={persoSelected ? `Polaire personnalisée active (${persoDetail})` : "Changer le type de bateau"}
-      >
-        {label}
-        <svg width="12" height="12" viewBox="0 0 12 12" fill="currentColor" style={{ opacity: 0.5 }}>
-          <path d="M2 4l4 4 4-4" stroke="currentColor" strokeWidth="1.5" fill="none" strokeLinecap="round"/>
-        </svg>
-      </button>
-
-      {open && (
-        <>
-          {/* backdrop */}
-          <div className="fixed inset-0 z-10" onClick={() => setOpen(false)} />
-          <div
-            className="absolute left-0 top-7 z-20 min-w-[240px] rounded-xl shadow-xl border overflow-hidden"
-            style={{ background: "var(--ow-bg-1)", borderColor: "var(--ow-line-2)", boxShadow: "var(--ow-shadow-pop)" }}
-          >
-            {persoDefined && (
-              <>
-                <button
-                  onClick={() => { onPersoSelect(); setOpen(false); }}
-                  className="w-full text-left px-4 py-3 text-sm transition-colors"
-                  style={{
-                    background: persoSelected ? "var(--ow-accent-soft)" : "transparent",
-                    color: persoSelected ? "var(--ow-accent)" : "var(--ow-fg-0)",
-                  }}
-                >
-                  <div className="font-semibold">Perso</div>
-                  <div className="text-[11px] mt-0.5" style={{ color: "var(--ow-fg-2)" }}>
-                    {persoDetail}
-                  </div>
-                </button>
-                <div
-                  className="flex gap-2 px-4 pb-2.5 text-[11px]"
-                  style={{
-                    background: persoSelected ? "var(--ow-accent-soft)" : "transparent",
-                    color: "var(--ow-accent)",
-                    borderBottom: "1px solid var(--ow-line)",
-                  }}
-                >
-                  <a href="/config" onClick={rememberReturnPath} className="underline">
-                    éditer
-                  </a>
-                  <span style={{ opacity: 0.4 }}>·</span>
-                  <button type="button" onClick={resetCustom} className="underline">
-                    réinitialiser
-                  </button>
-                </div>
-              </>
-            )}
-            {archetypes.map((a) => (
-              <button
-                key={a.slug}
-                onClick={() => { onChange(a.slug); setOpen(false); }}
-                className="w-full text-left px-4 py-3 text-sm transition-colors"
-                style={{
-                  background: !persoSelected && a.slug === currentSlug ? "var(--ow-accent-soft)" : "transparent",
-                  color: !persoSelected && a.slug === currentSlug ? "var(--ow-accent)" : "var(--ow-fg-0)",
-                  borderBottom: "1px solid var(--ow-line)",
-                }}
-              >
-                <div className="font-semibold">{a.name}</div>
-                <div className="text-[11px] mt-0.5" style={{ color: "var(--ow-fg-2)" }}>
-                  {a.length_ft} ft · {a.type} · {a.examples[0]}
-                </div>
-              </button>
-            ))}
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
-
-// ── LegList ──────────────────────────────────────────────────────────────────
-// Click-to-expand list of legs. Collapsed = single natural-language summary
-// line ("Tronçon 1 : 45 mn au près avec mer formée"). Expanded = a 4-block KPI
-// grid (vent / mer / distance / temps) above the existing compass-and-build-up
-// LegDetailCard so the user can scan or drill.
-
-function fmtHM(iso: string): string {
-  return new Date(iso).toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" });
-}
-
-function SummaryCell({ value }: { value: string | null }) {
-  if (!value) return null;
-  return (
-    <span
-      className="text-xs whitespace-normal"
-      style={{ color: "var(--ow-fg-1)", lineHeight: 1.15, display: "inline-block" }}
-    >
-      {value}
-    </span>
-  );
-}
-
-function KpiBlock({
-  value,
-  label,
-  tone,
-}: {
-  value: string;
-  label?: string;
-  tone?: "default" | "warn";
-}) {
-  return (
-    <div
-      className="rounded-md border px-2 py-1"
-      style={{
-        background: tone === "warn"
-          ? "color-mix(in srgb, var(--ow-warn, #fbbf24) 14%, transparent)"
-          : "var(--ow-bg-1)",
-        borderColor: tone === "warn"
-          ? "color-mix(in srgb, var(--ow-warn, #fbbf24) 38%, transparent)"
-          : "var(--ow-line)",
-      }}
-    >
-      <div
-        className="text-[11px] font-semibold tabular-nums leading-tight break-words"
-        style={{ color: "var(--ow-fg-0)", fontFamily: "var(--ow-font-mono)" }}
-      >
-        {value}
-      </div>
-      {label && (
-        <div
-          className="text-[9px] uppercase tracking-wider leading-tight mt-0.5 break-words"
-          style={{ color: "var(--ow-fg-2)" }}
-        >
-          {label}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function LegRow({
-  leg,
-  index,
-  expanded,
-  onToggle,
-}: {
-  leg: AggregatedLeg;
-  index: number;
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  const cx = cxLevel((leg.tws_min + leg.tws_max) / 2);
-  const summary = buildLegSummaryCells(leg);
-
-  // KPI values shown on expand. Wind + allure are intentionally absent —
-  // the collapsed row already carries them, no point repeating.
-  // Compact French formatting: "1,8m (6s)" matches sailing-French copy.
-  const fr1 = (n: number) => n.toFixed(1).replace(".", ",");
-
-  const seaValue = leg.hs_avg_m == null
-    ? "—"
-    : leg.tp_avg_s != null
-      ? `${fr1(leg.hs_avg_m)}m (${leg.tp_avg_s.toFixed(0)}s)`
-      : `${fr1(leg.hs_avg_m)}m`;
-  const seaLabel = leg.hs_avg_m == null
-    ? "mer non observée"
-    : leg.sea_direction === "face"
-      ? "de face"
-      : leg.sea_direction === "travers"
-        ? "de travers"
-        : "par l'arrière";
-
-  // Warn tint when sea state notable. Mirrors the same Hs threshold the
-  // summary line uses, so "Mer Formée" badge and warn-coloured KPI agree.
-  const seaWarn = leg.hs_avg_m != null && leg.hs_avg_m > 1.25;
-
-  const handleKey = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter" || e.key === " ") {
-      e.preventDefault();
-      onToggle();
-    }
-  };
-  const rowBg = expanded ? "var(--ow-bg-2)" : "transparent";
-
-  // Each leg returns three `<tr>`s into the shared `<table>` in LegList:
-  // a title row (badge + name + speed + chevron), a chip row (the four
-  // summary cells), and an optional expand row (KPIs + LegDetailCard).
-  // Because they're all in the same table, the colgroup defined in LegList
-  // forces every leg's chip cells to live in the same column widths —
-  // exactly the cross-row alignment a tableless flex/grid layout couldn't
-  // give us when each LegRow had its own grid container.
-  return (
-    <>
-      {/* Single-row leg: badge + 4 info cells + chevron. The speed
-          indicator was dropped and the redundant "Tronçon X" label was
-          dropped earlier — the numbered badge identifies the leg. */}
-      <tr
-        role="button"
-        tabIndex={0}
-        aria-expanded={expanded}
-        onClick={onToggle}
-        onKeyDown={handleKey}
-        className="cursor-pointer"
-        style={{ background: rowBg }}
-      >
-        <td className="py-2 pl-3 pr-2 align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          {/* Leg label is "from→to" using user-waypoint indices (1-based).
-              `index` here is 0-based across legs, so leg N goes from
-              waypoint N to waypoint N+1. Distance sits under the badge so
-              the user can scan leg length without expanding. */}
-          <div className="flex flex-col items-start gap-0.5">
-            <span
-              className="inline-flex h-6 px-1.5 rounded-md items-center justify-center text-[10px] font-bold tabular-nums whitespace-nowrap"
-              style={{ background: CX_COLORS[cx], color: "#0B1D14", fontFamily: "var(--ow-font-mono)" }}
-            >
-              {index + 1}→{index + 2}
-            </span>
-            <span
-              className="text-[10px] tabular-nums whitespace-nowrap"
-              style={{ color: "var(--ow-fg-2)", fontFamily: "var(--ow-font-mono)" }}
-            >
-              {fr1(leg.distance_nm)} nm
-            </span>
-          </div>
-        </td>
-        <td className="py-2 px-1 align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          <SummaryCell value={summary.duration} />
-        </td>
-        <td className="py-2 px-1 align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          <SummaryCell value={summary.allure} />
-        </td>
-        <td className="py-2 px-1 align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          <SummaryCell value={summary.wind} />
-        </td>
-        <td className="py-2 px-1 align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          <SummaryCell value={summary.flag} />
-        </td>
-        <td className="py-2 pl-1 pr-3 text-right align-middle" style={{ borderTop: "1px solid var(--ow-line)" }}>
-          <span
-            aria-hidden="true"
-            className="inline-flex"
-            style={{
-              color: expanded ? "var(--ow-accent)" : "var(--ow-fg-3)",
-              transform: expanded ? "rotate(180deg)" : "none",
-              transition: "transform 150ms ease, color 150ms ease",
-            }}
-          >
-            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <path d="M3 6l5 5 5-5" />
-            </svg>
-          </span>
-        </td>
-      </tr>
-      {expanded && (
-        <tr style={{ background: rowBg }}>
-          <td colSpan={6} className="px-4 pb-3">
-            {/* Two KPI cells (time / sea). Wind, allure and distance all
-                appear in the collapsed row above; repeating them in the
-                expand was visual noise. */}
-            <div className="grid grid-cols-2 gap-2 mb-3">
-              <KpiBlock value={`${fmtHM(leg.start_time)} → ${fmtHM(leg.end_time)}`} label="dep → arr" />
-              <KpiBlock value={seaValue} label={seaLabel} tone={seaWarn ? "warn" : "default"} />
-            </div>
-            <LegDetailCard leg={leg} />
-          </td>
-        </tr>
-      )}
-    </>
-  );
-}
-
-function LegList({
-  legs,
-  openIdx,
-  onOpenChange,
-}: {
-  legs: AggregatedLeg[];
-  openIdx: number | null;
-  onOpenChange: (idx: number | null) => void;
-}) {
-  return (
-    <div>
-      <table className="w-full" style={{ borderCollapse: "collapse" }}>
-        <thead>
-          <tr
-            className="text-[9px] uppercase tracking-widest"
-            style={{ color: "var(--ow-fg-3)" }}
-          >
-            <th className="py-1 pl-3 pr-2 pt-3 text-left font-semibold">Tronçon</th>
-            <th className="py-1 px-1 pt-3 text-left font-semibold">Durée</th>
-            <th className="py-1 px-1 pt-3 text-left font-semibold">Allure</th>
-            <th className="py-1 px-1 pt-3 text-left font-semibold">Vent (kn)</th>
-            <th className="py-1 px-1 pt-3 text-left font-semibold">Mer</th>
-            <th className="py-1 pl-1 pr-3 pt-3" />
-          </tr>
-        </thead>
-        <tbody>
-          {legs.map((leg, i) => (
-            <LegRow
-              key={i}
-              leg={leg}
-              index={i}
-              expanded={openIdx === i}
-              onToggle={() => onOpenChange(openIdx === i ? null : i)}
-            />
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-}
-
-// ── PlanSidebar ───────────────────────────────────────────────────────────────
-
-interface PlanSidebarProps {
-  passage: PassageReport | null;
-  complexity: ComplexityScore | null;
-  isLoading: boolean;
-  error: string | null;
-  archetypes: Archetype[];
-  currentArchetypeSlug: string;
-  onArchetypeChange: (slug: string) => void;
-  /** Select the perso polar (first entry of the boat list when defined). */
-  onPersoSelect: () => void;
-  departure: string;
-  onDepartureChange: (iso: string) => void;
-  isStale: boolean;
-  onRefetch: () => void;
-  forecastUpdatedAt: string | null;
-  waypointCount: number;
-  waypoints: [number, number][];
-  // ETA-driven sub-mode for "Simuler ma route".
-  timeAnchor: TimeAnchor;
-  onTimeAnchorChange: (a: TimeAnchor) => void;
-  // Compare-windows mode (lifted from local state in step 2)
-  mode: PlanMode;
-  onModeChange: (m: PlanMode) => void;
-  sweepEarliest: string;
-  sweepLatest: string;
-  sweepIntervalHours: number;
-  onSweepEarliestChange: (iso: string) => void;
-  onSweepLatestChange: (iso: string) => void;
-  onSweepIntervalChange: (h: number) => void;
-  windows: PassageWindow[] | null;
-  metaWarnings: string[];
-  onCompareFetch: () => void;
-  onWindowSelect?: (w: PassageWindow) => void;
-  /** Selected leg index in the filled view (drives the map highlight). */
-  selectedLegIdx: number | null;
-  onSelectedLegChange: (idx: number | null) => void;
-  /** Discard the current plan: clears waypoints, results, cache, URL. */
-  onReset?: () => void;
-  /** True once the user has interacted with the mode pills after placing
-   *  2+ waypoints. While false, the sidebar collapses to just the toggle +
-   *  reset button (compact "pick a mode" view). Set on first pill click. */
-  actionTaken?: boolean;
-}
-
-// Header row: ModeToggle + an optional trash button to discard the plan.
-// The trash sits flush with the toggle and only renders when a route exists
-// (no waypoints → nothing to reset). Tooltip: "Réinitialiser".
-function PlanHeaderRow({
-  mode,
-  onModeChange,
-  locked,
-  pristine,
-  onReset,
-}: {
-  mode: PlanMode;
-  onModeChange: (m: PlanMode) => void;
-  locked?: boolean;
-  pristine?: boolean;
-  onReset?: () => void;
-}) {
-  return (
-    <div className="flex items-stretch gap-2">
-      <div className="flex-1 min-w-0">
-        <ModeToggle value={mode} onChange={onModeChange} locked={locked} pristine={pristine} />
-      </div>
-      {onReset && (
-        <button
-          type="button"
-          onClick={onReset}
-          title="Nouveau plan"
-          aria-label="Nouveau plan"
-          className="shrink-0 flex items-center justify-center rounded-lg transition-colors hover:opacity-100"
-          style={{
-            width: 38,
-            background: "var(--ow-bg-2)",
-            border: "1px solid var(--ow-line)",
-            color: "var(--ow-fg-2)",
-            opacity: 0.85,
-          }}
-        >
-          <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <path d="M2.5 4h11" />
-            <path d="M6 4V2.5h4V4" />
-            <path d="M3.5 4l.9 9.2a1 1 0 0 0 1 .8h5.2a1 1 0 0 0 1-.8L12.5 4" />
-            <path d="M6.5 6.5v5" />
-            <path d="M9.5 6.5v5" />
-          </svg>
-        </button>
-      )}
-    </div>
-  );
-}
-
-export function PlanSidebar({
-  passage,
-  complexity,
-  isLoading,
-  error,
-  archetypes,
-  currentArchetypeSlug,
-  onArchetypeChange,
-  onPersoSelect,
-  departure,
-  onDepartureChange,
-  isStale,
-  onRefetch,
-  forecastUpdatedAt,
-  waypointCount,
-  waypoints,
-  timeAnchor,
-  onTimeAnchorChange,
-  mode,
-  onModeChange,
-  sweepEarliest,
-  sweepLatest,
-  sweepIntervalHours,
-  onSweepEarliestChange,
-  onSweepLatestChange,
-  onSweepIntervalChange,
-  windows,
-  metaWarnings,
-  onCompareFetch,
-  onWindowSelect,
-  selectedLegIdx,
-  onSelectedLegChange,
-  onReset,
-  actionTaken = true,
-}: PlanSidebarProps) {
-  // Show the trash button only when the user has placed enough to have
-  // something to clear — nothing to reset on a fully-empty form.
-  const resetHandler = onReset && waypointCount >= 2 ? onReset : undefined;
-  const { resolvedTheme } = useTheme();
-  const sweepValid = mode === "compare"
-    ? validateSweep(sweepEarliest, sweepLatest, sweepIntervalHours)
-    : { ok: true } as SweepValidation;
-  const canCalculate = waypointCount >= 2 && (mode === "single" || sweepValid.ok);
-  const [isEditingParams, setIsEditingParams] = useState(false);
-
-  // Mode selection now lives entirely in the PlanHeaderRow pills at the top
-  // of the sidebar (which always show both options and let the user toggle
-  // freely). The intermediate "Que voulez-vous faire ?" picker that used to
-  // appear once after the second waypoint was redundant once the toggle
-  // became permanent, so it was removed.
-  const handleModeChange = onModeChange;
-
-  if (isLoading) {
-    return (
-      <div className="p-4 space-y-4 animate-fade-in">
-        <div className="skeleton h-8 w-48 rounded-lg" />
-        <div className="skeleton h-5 w-32 rounded" />
-        <div className="grid grid-cols-4 gap-2">
-          {[0, 1, 2, 3].map((i) => <div key={i} className="skeleton h-14 rounded-lg" />)}
-        </div>
-        <div className="skeleton h-2.5 rounded-full" />
-        {[0, 1, 2].map((i) => <div key={i} className="skeleton h-10 rounded-lg" />)}
-      </div>
-    );
-  }
-
-  if (error) {
+  // 2. failed
+  if (apiError) {
     return (
       <div className="p-4">
-        <PlanHeaderRow mode={mode} onModeChange={handleModeChange} locked={waypointCount < 2} onReset={resetHandler} />
+        <PlanHeaderRow locked={waypoints.length < 2} />
         <div className="mt-4 rounded-xl p-4 text-sm" style={{ background: "var(--ow-err-soft)", color: "var(--ow-err)", border: "1px solid var(--ow-err-line)" }}>
           <p className="font-semibold mb-1">Erreur</p>
-          <p className="leading-relaxed">{error}</p>
+          <p className="leading-relaxed">{apiError}</p>
         </div>
       </div>
     );
   }
 
-  // ── Empty state: no waypoints yet ─────────────────────────────────────────
-  if (waypointCount < 2) {
+  // 3. no route yet
+  if (waypoints.length < 2) {
     return (
       <div className="p-4 animate-fade-in">
-        <PlanHeaderRow mode={mode} onModeChange={handleModeChange} locked />
+        <PlanHeaderRow locked />
         <EmptyState />
       </div>
     );
   }
 
-  // ── Compact picker: 2+ waypoints but user hasn't picked a mode yet.
-  //    Mobile: pills + trash only (vertical real-estate is precious — the
-  //    drawer animates up just enough for the toggle).
-  //    Desktop: pills + trash + the larger "narrative" cards underneath,
-  //    which reassure first-time users with example phrasings.
-  //    In both cases, clicking any pill or card flips ``actionTaken``
-  //    upstream and unlocks the full sidebar.
+  // 4. a route, no mode picked.
+  // Mobile: pills and trash only, vertical real estate is precious and the
+  // drawer only animates up enough for the toggle. Desktop adds the larger
+  // narrative cards, which reassure first-time users with example phrasings.
+  // Either way, clicking a pill or a card unlocks the full panel.
   if (!actionTaken) {
     return (
       <div className="p-4 space-y-4 animate-fade-in">
-        <PlanHeaderRow
-          mode={mode}
-          onModeChange={handleModeChange}
-          pristine
-          onReset={resetHandler}
-        />
+        <PlanHeaderRow pristine />
         <div className="hidden lg:block">
-          <ModePicker onPick={(m) => handleModeChange(m)} />
+          <ModePicker onPick={actions.setMode} />
         </div>
       </div>
     );
   }
 
-  const archetype = archetypes.find((a) => a.slug === currentArchetypeSlug);
-  // Recap-strip boat label. Perso active: the polar's provenance — the file
-  // name suffixed « importée », or cfg.base suffixed « ajustée » (the hull
-  // the tuning was built on and the one the ArchetypeSelector announces —
-  // never the page slug, which could still carry another boat from a stale
-  // URL/cache, #220). Otherwise the French display label of the selected
-  // stock archetype. Re-read the config at render, same rationale as
-  // ArchetypeSelector. The performance coefficient is deliberately not
-  // repeated here: /config's Bateau tab stays its single indicator.
-  const recapPolarCfg = loadPolarConfig();
-  const importedName = recapPolarCfg.imported?.name ?? "polaire";
-  const archetypeLabel = isPersoActive(recapPolarCfg)
-    ? isImportedActive(recapPolarCfg)
-      ? `${importedName.length > 20 ? `${importedName.slice(0, 19)}…` : importedName} (importée)`
-      : `${ARCHETYPE_LABELS[recapPolarCfg.base] ?? recapPolarCfg.base} (ajustée)`
-    : ARCHETYPE_LABELS[currentArchetypeSlug] ?? archetype?.name ?? currentArchetypeSlug;
-
-  // ── Filled compare view: results loaded, inputs collapsed into a recap ────
+  // 5. the picked mode.
+  const label = boatLabel(polarConfig, archetype, archetypes);
   if (mode === "compare" && windows && windows.length > 0) {
-    const fmtShort = (iso: string) =>
-      new Date(iso).toLocaleDateString("fr-FR", { weekday: "short", day: "numeric", month: "short" });
-    const fmtClock = (iso: string) =>
-      new Date(iso).toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" });
-    const recapPrimary = `${fmtShort(sweepEarliest)} ${fmtClock(sweepEarliest)} → ${fmtShort(sweepLatest)} ${fmtClock(sweepLatest)}`;
-    const recapSecondary = `pas ${sweepIntervalHours}h · ${archetypeLabel}`;
-
-    return (
-      <div className="animate-fade-in">
-        <div className="px-4 pt-4 pb-3" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-          <PlanHeaderRow mode={mode} onModeChange={handleModeChange} onReset={resetHandler} />
-        </div>
-
-        <div className="px-4 py-2.5" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-          <button
-            onClick={onCompareFetch}
-            disabled={!canCalculate}
-            className="w-full flex items-center justify-center gap-2 rounded-md py-1.5 text-xs font-semibold transition-all"
-            style={{
-              background: canCalculate ? "#F4C25C" : "var(--ow-bg-2)",
-              color: canCalculate ? "#3a2a08" : "var(--ow-fg-3)",
-              border: `1px solid ${canCalculate ? "transparent" : "var(--ow-line)"}`,
-            }}
-          >
-            <svg width="11" height="11" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <path d="M13.5 2.5A7 7 0 1 0 14.5 9" /><path d="M14 1v4h-4" />
-            </svg>
-            Recalculer
-          </button>
-        </div>
-
-        {/* Zero-height marker: when results land, the mobile drawer jumps
-            here so the recap + windows table open the view and the pills +
-            Recalculer block above stays one scroll-up away. */}
-        <div data-results-anchor />
-        <RecapButton
-          primary={recapPrimary}
-          secondary={recapSecondary}
-          isOpen={isEditingParams}
-          onClick={() => setIsEditingParams((v) => !v)}
-        />
-        {isEditingParams && (
-          <div className="px-4 py-3 space-y-3" style={{ borderBottom: "1px solid var(--ow-line)", background: "var(--ow-bg-2)" }}>
-            <SweepForm
-              earliest={sweepEarliest}
-              latest={sweepLatest}
-              intervalHours={sweepIntervalHours}
-              onEarliestChange={onSweepEarliestChange}
-              onLatestChange={onSweepLatestChange}
-              onIntervalChange={onSweepIntervalChange}
-            />
-            <ArchetypeSelector
-              currentSlug={currentArchetypeSlug}
-              archetypes={archetypes}
-              onChange={onArchetypeChange}
-              onPersoSelect={onPersoSelect}
-              onCustomCleared={onRefetch}
-            />
-          </div>
-        )}
-
-        {metaWarnings.length > 0 && (
-          <div className="px-4 py-2.5 space-y-1.5" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-            {metaWarnings.map((m, i) => <Warn key={i}>{m}</Warn>)}
-          </div>
-        )}
-
-        {/* When the route has been edited since the sweep, the cached windows
-            (and the per-window passages they'd drill into) describe the *old*
-            itinerary — opening one would render a route that no longer matches
-            the map. Hide the table behind a recompute prompt, mirroring single
-            mode (#152). */}
-        {isStale ? (
-          <div
-            className="px-4 py-6 text-center text-xs"
-            style={{ color: "var(--ow-fg-2)", borderBottom: "1px solid var(--ow-line)" }}
-          >
-            Itinéraire modifié. Cliquez sur Recalculer pour comparer les créneaux du nouveau trajet.
-          </div>
-        ) : (
-          <>
-            <WindowsTable windows={windows} onSelect={onWindowSelect} />
-
-            <p className="px-4 py-2 text-[10px]" style={{ color: "var(--ow-fg-3)", borderTop: "1px solid var(--ow-line)" }}>
-              {windows.length} fenêtre{windows.length > 1 ? "s" : ""} comparée{windows.length > 1 ? "s" : ""} · cliquez sur une ligne pour ouvrir la simulation détaillée
-            </p>
-          </>
-        )}
-      </div>
-    );
+    return <CompareResults windows={windows} boatLabel={label} canCalculate={canCalculate} />;
   }
-
-  // ── Form state (compare without windows; single before any result) ────────
   if (mode === "compare" || !passage || !complexity) {
-    const accent = mode === "compare" ? "#F4C25C" : "var(--ow-accent)";
-    const ctaInk = mode === "compare" ? "#3a2a08" : "#fff";
-    return (
-      <div className="p-4 space-y-3 animate-fade-in">
-        <PlanHeaderRow mode={mode} onModeChange={handleModeChange} onReset={resetHandler} />
-
-        {/* Tab-panel: everything below is "the contents" of the picked mode.
-            The 2 px accent stripe at the top picks up the mode's color
-            (cyan for Simuler, amber for Comparer) so the panel reads as the
-            payload of the active pill above. */}
-        <div
-          className="rounded-xl p-4 space-y-3"
-          style={{
-            background: "var(--ow-bg-1)",
-            border: "1px solid var(--ow-line)",
-            borderTop: `2px solid ${accent}`,
-          }}
-        >
-          {mode === "single" ? (
-            <div className="space-y-3">
-              <TimeAnchorToggle value={timeAnchor} onChange={onTimeAnchorChange} />
-              <DepartureSlider
-                value={departure}
-                onChange={onDepartureChange}
-                resolvedTheme={resolvedTheme}
-                anchor={timeAnchor}
-              />
-            </div>
-          ) : (
-            <SweepForm
-              earliest={sweepEarliest}
-              latest={sweepLatest}
-              intervalHours={sweepIntervalHours}
-              onEarliestChange={onSweepEarliestChange}
-              onLatestChange={onSweepLatestChange}
-              onIntervalChange={onSweepIntervalChange}
-            />
-          )}
-
-          <ArchetypeSelector
-            currentSlug={currentArchetypeSlug}
-            archetypes={archetypes}
-            onChange={onArchetypeChange}
-            onPersoSelect={onPersoSelect}
-            onCustomCleared={onRefetch}
-          />
-
-          <button
-            onClick={mode === "single" ? onRefetch : onCompareFetch}
-            disabled={!canCalculate}
-            className="w-full flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl text-sm font-bold transition-all"
-            style={{
-              background: canCalculate ? accent : "var(--ow-bg-2)",
-              color: canCalculate ? ctaInk : "var(--ow-fg-3)",
-              border: `1px solid ${canCalculate ? "transparent" : "var(--ow-line-2)"}`,
-              cursor: canCalculate ? "pointer" : "not-allowed",
-            }}
-          >
-            <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <path d="M13.5 2.5A7 7 0 1 0 14.5 9"/><path d="M14 1v4h-4"/>
-            </svg>
-            {canCalculate
-              ? mode === "single" ? "Calculer le passage" : "Comparer les créneaux"
-              : `${waypointCount}/2 waypoints`}
-          </button>
-        </div>
-      </div>
-    );
+    return <PlanForm canCalculate={canCalculate} />;
   }
-
-  // ── Filled state: single mode with a result ───────────────────────────────
-  const hasWarnings = (complexity.warnings?.length ?? 0) > 0 || passage.warnings.length > 0;
-  const recapDate = new Date(departure).toLocaleDateString("fr-FR", { weekday: "short", day: "numeric", month: "short" });
-  const recapTime = new Date(departure).toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" });
-
-  return (
-    <div className="animate-fade-in">
-      {/* Mode tabs */}
-      <div className="px-4 pt-4 pb-3" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-        <PlanHeaderRow mode={mode} onModeChange={handleModeChange} onReset={resetHandler} />
-      </div>
-
-      {/* Recalculer bar */}
-      <div className="px-4 py-2.5" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-        <button
-          onClick={onRefetch}
-          className="w-full flex items-center justify-center gap-2 rounded-md py-1.5 text-xs font-semibold transition-all"
-          style={{
-            background: isStale ? "var(--ow-accent)" : "var(--ow-bg-2)",
-            color: isStale ? "#fff" : "var(--ow-fg-1)",
-            border: `1px solid ${isStale ? "transparent" : "var(--ow-line)"}`,
-          }}
-        >
-          <svg width="11" height="11" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-            <path d="M13.5 2.5A7 7 0 1 0 14.5 9" /><path d="M14 1v4h-4" />
-          </svg>
-          {isStale ? "Recalculer" : "Recalculer"}
-        </button>
-      </div>
-
-      {/* Récap compact: click to edit departure / archetype inline.
-          The zero-height marker above it is where the mobile drawer jumps
-          when results land — recap + legs open the view, pills + Recalculer
-          stay one scroll-up away. */}
-      <div data-results-anchor />
-      <RecapButton
-        primary={`${timeAnchor === "arrival" ? "Arrivée" : "Départ"} : ${recapDate.charAt(0).toUpperCase() + recapDate.slice(1)} · ${recapTime}`}
-        secondary={archetypeLabel}
-        isOpen={isEditingParams}
-        onClick={() => setIsEditingParams((v) => !v)}
-      />
-      {isEditingParams && (
-        <div className="px-4 py-3 space-y-3" style={{ borderBottom: "1px solid var(--ow-line)", background: "var(--ow-bg-2)" }}>
-          <TimeAnchorToggle value={timeAnchor} onChange={onTimeAnchorChange} />
-          <DepartureSlider
-            value={departure}
-            onChange={onDepartureChange}
-            resolvedTheme={resolvedTheme}
-            anchor={timeAnchor}
-          />
-          <ArchetypeSelector
-            currentSlug={currentArchetypeSlug}
-            archetypes={archetypes}
-            onChange={onArchetypeChange}
-            onPersoSelect={onPersoSelect}
-            onCustomCleared={onRefetch}
-          />
-        </div>
-      )}
-
-      {/* Total route stats (Distance / Durée / Arrivée + segment bar).
-          Desktop only: on mobile the floating overlay (PlanHeroStats) stays
-          the single source of truth for these totals, per the b90a5bf
-          decision. Hidden entirely when the route was edited without
-          recalculating, same rule as the mobile overlay. */}
-      {!isStale && (
-        <div className="hidden lg:block px-4 py-3.5" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-          <HeroStats passage={passage} />
-        </div>
-      )}
-
-      {/* Warnings */}
-      {hasWarnings && (
-        <div className="px-4 py-2.5 space-y-1.5" style={{ borderBottom: "1px solid var(--ow-line)" }}>
-          {complexity.warnings?.map((w, i) => <Warn key={i}>{w.message}</Warn>)}
-          {passage.warnings.map((w, i) => <Warn key={`pw-${i}`}>{w}</Warn>)}
-        </div>
-      )}
-
-      {/* Legs — click any row to see the build-up.
-          Hidden while the plan is stale: re-mapping old segments to a freshly
-          edited waypoint list is unsafe (cf. aggregateLegs's index juggling)
-          and a silent gray-out could be missed on a quick glance. Showing a
-          placeholder is plainer and forces a recompute before the user reads
-          numbers that no longer match the route on the map. */}
-      {isStale ? (
-        <div
-          className="px-4 py-6 text-center text-xs"
-          style={{ color: "var(--ow-fg-2)", borderBottom: "1px solid var(--ow-line)" }}
-        >
-          Itinéraire modifié. Cliquez sur Recalculer pour mettre à jour les détails.
-        </div>
-      ) : (
-        (() => {
-          // Min upwind angle of the boat this plan was computed for, so a
-          // leg whose direct course sits in the no-go zone reads
-          // "Près (louvoyage)" instead of a close-hauled label the boat
-          // cannot hold (#277).
-          const legs = aggregateLegs(
-            passage.segments,
-            waypoints,
-            passage.efficiency,
-            planMinUpwind(recapPolarCfg, currentArchetypeSlug),
-          );
-          return (
-            <LegList
-              legs={legs}
-              openIdx={selectedLegIdx}
-              onOpenChange={onSelectedLegChange}
-            />
-          );
-        })()
-      )}
-
-      {/* Footer */}
-      {forecastUpdatedAt && (
-        <p className="px-4 py-2 text-[10px]" style={{ color: "var(--ow-fg-2)", borderTop: "1px solid var(--ow-line)" }}>
-          Données fraîches au {new Date(forecastUpdatedAt).toLocaleTimeString("fr-FR", { hour: "2-digit", minute: "2-digit" })} · Open-Meteo.com (CC BY 4.0)
-        </p>
-      )}
-    </div>
-  );
+  return <SingleResults passage={passage} complexity={complexity} boatLabel={label} />;
 }
